+++ core/oputil/oputil_file_base.js
@@ -209,7 +209,7 @@ function scanFileAreaForChanges(areaInfo, options, cb) {
                                 async.series(
                                     [
                                         function quickCheck(next) {
-                                            if(options['full-scan']) {
+                                            if(options['full']) {
                                                 return next(null);
                                             }
 
@@ -228,6 +228,16 @@ function scanFileAreaForChanges(areaInfo, options, cb) {
                                                 {
                                                     areaTag		: areaInfo.areaTag,
                                                     storageTag	: storageLoc.storageTag
+                                                },
+                                                (stepInfo, next) => {
+                                                    if(argv.verbose) {
+                                                        if(stepInfo.error) {
+                                                            console.error(`  error: ${stepInfo.error}`);
+                                                        } else {
+                                                            console.info(`  processing: ${stepInfo.step}`);
+                                                        }
+                                                    }
+                                                    return next(null);
                                                 },
                                                 (err, fileEntry, dupeEntries) => {
                                                     if(err) {
@@ -477,7 +487,7 @@ function scanFileAreas() {
     }
 
     options.descFile 	    = argv['desc-file'];	//	--desc-file or --desc-file PATH
-    options['full-scan'] 	= argv['-full-scan'];
+    options['full'] 	    = argv.full;
 
     options.areaAndStorageInfo = getAreaAndStorage(argv._.slice(2));